+++ mobile/android/android-components/components/lib/crash/src/main/java/mozilla/components/lib/crash/service/SentryService.kt
@@ -9,10 +9,14 @@ import android.net.Uri
 import io.sentry.SentryClient
 import io.sentry.SentryClientFactory
 import io.sentry.android.AndroidSentryClientFactory
+import mozilla.components.Build
 import mozilla.components.lib.crash.Crash
 
 /**
  * A [CrashReporterService] implementation that uploads crash reports to a Sentry server.
+ *
+ * This implementation will add default tags to every sent crash report (like the used Android Components version)
+ * prefixed with "ac.".
  *
  * @param context The application [Context].
  * @param dsn Data Source Name of the Sentry server.
@@ -34,6 +38,11 @@ class SentryService(
         tags.forEach { entry ->
             addTag(entry.key, entry.value)
         }
+
+        // Add default tags
+        addTag("ac.version", Build.version)
+        addTag("ac.git", Build.gitHash)
+        addTag("ac.as.build_version", Build.applicationServicesVersion)
     }
 
     override fun report(crash: Crash.UncaughtExceptionCrash) {