+++ mobile/android/android-components/docs/changelog.md
@@ -33,6 +33,9 @@ permalink: /changelog/
 * **support-base**
   * Added `Build` object that contains information about the current Android Components build (like version number and git hash).
 
+* **lib-crash**
+  * Crash reports sent to Sentry now contain additional tags about the used Android Components version and setup (prefixed with "ac.").
+
 # 0.53.0
 
 * [Commits](https://github.com/mozilla-mobile/android-components/compare/v0.52.0...v0.53.0)